rename sql mixin file
